refactor: add-infrastructure and add-application moved to respective projects
