+++ .config/dotnet-tools.json
@@ -3,7 +3,7 @@
   "isRoot": true,
   "tools": {
     "dotnet-ef": {
-      "version": "8.0.1",
+      "version": "8.0.4",
       "commands": [
         "dotnet-ef"
       ]
