+++ src/Application/AddApplication.cs
@@ -1,16 +1,12 @@
-using System.Reflection;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Server.AddServices;
+namespace Application;
 
 public static class AddApplication
 {
     public static IServiceCollection AddApplicationServices(this IServiceCollection services)
     {
-        services.AddMediatR(cfg =>
-        {
-            cfg.RegisterServicesFromAssembly(typeof(Application.AssemblyName).Assembly);
-        });
+        services.AddMediatR(cfg => { cfg.RegisterServicesFromAssembly(typeof(Application.AssemblyName).Assembly); });
         return services;
     }
 }
