+++ src/Application/Interfaces/IApplicationDbContext.cs
@@ -1,6 +1,5 @@
 using System.Threading;
 using System.Threading.Tasks;
-using Domain;
 using Domain.FoodItems;
 using Microsoft.EntityFrameworkCore;
