+++ src/Client/package.json
@@ -1,7 +1,7 @@
 {
     "name": "nutrition-tracker-client",
     "private": true,
-    "version": "0.10.0",
+    "version": "0.11.0",
     "author": {
         "name": "Vidar André Bø"
     },
@@ -40,7 +40,7 @@
         "rimraf": "^5.0.5",
         "tailwindcss": "^3.4.1",
         "typescript": "^5.4.2",
-        "vite": "^5.1.5",
+        "vite": "^5.2.8",
         "vitest": "^1.3.1"
     }
 }
