+++ src/Domain/FoodItems/FoodItemDTO.cs
@@ -2,5 +2,10 @@
 
 namespace Domain.FoodItems
 {
-    public record FoodItemDTO(Guid Id, string Brand, string ProductName, NutritionalContent NutritionalContent, Guid OwnerId);
+    public record FoodItemDTO(
+        Guid Id,
+        string Brand,
+        string ProductName,
+        NutritionalContent NutritionalContent,
+        Guid OwnerId);
 }
