+++ src/Infrastructure/ManageMigrations.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Threading.Tasks;
-using Infrastructure;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
@@ -8,7 +7,7 @@ using Microsoft.Extensions.Hosting;
 using Npgsql;
 using Serilog;
 
-namespace Server.AddServices;
+namespace Infrastructure;
 
 public static class ManageMigrations
 {
