+++ src/Server/appsettings.json
@@ -5,7 +5,7 @@
       "Microsoft.AspNetCore": "Warning"
     }
   },
-  "Urls":"http://*:80",
+  "Urls": "http://*:80",
   "AllowedHosts": "*",
   "Serilog": {
     "MinimumLevel": {
